v1.4 (First Complete Release) Library and Samples
This update adds the Internet Shield (and it's associated modules) with
2 Samples.  Also all instance and callback support is now in-place
(similar to the way the Arduino Libraries handle things).

+++ OneSheeld/AccelerometerSensorShield.cs
@@ -6,7 +6,7 @@ namespace OneSheeldClasses
     public class AccelerometerSensorShield : XYZSensorShield
     {
         public AccelerometerSensorShield(OneSheeld onesheeld) : 
-            base(onesheeld, ACCELEROMETER_VALUE, (byte) ShieldIds.ACCELEROMETER_ID) {}
+            base(onesheeld, ACCELEROMETER_VALUE, ShieldIds.ACCELEROMETER_ID) {}
 
         //Input Function ID 
         const byte ACCELEROMETER_VALUE = 0x01;
