+++ OneSheeld/GyroscopeSensorShield.cs
@@ -6,7 +6,7 @@ namespace OneSheeldClasses
     public class GyroscopeSensorShield : XYZSensorShield
     {
         public GyroscopeSensorShield(OneSheeld onesheeld) : 
-            base(onesheeld, GYROSCOPE_VALUE, (byte) ShieldIds.GYROSCOPE_ID) {}
+            base(onesheeld, GYROSCOPE_VALUE, ShieldIds.GYROSCOPE_ID) {}
 
         //Input Function ID 
         const byte GYROSCOPE_VALUE = 0x01;
