+++ OneSheeld/PressureSensorShield.cs
@@ -6,7 +6,7 @@ namespace OneSheeldClasses
     public class PressureSensorShield : ULongInputShield
     {
         public PressureSensorShield(OneSheeld onesheeld) :
-            base(onesheeld, PRESSURE_VALUE, (byte) ShieldIds.PRESSURE_ID, 2) { }
+            base(onesheeld, PRESSURE_VALUE, ShieldIds.PRESSURE_ID, 2) { }
 
         const byte PRESSURE_VALUE = 0x01;
     }
