+++ OneSheeld/ToggleButtonShield.cs
@@ -6,7 +6,7 @@ namespace OneSheeldClasses
     public class ToggleButtonShield : BoolInputShield
     {
         public ToggleButtonShield(OneSheeld onesheeld)
-            : base(onesheeld, TOGGLEBUTTON_VALUE, (byte) ShieldIds.TOGGLE_BUTTON_ID) { }
+            : base(onesheeld, TOGGLEBUTTON_VALUE, ShieldIds.TOGGLE_BUTTON_ID) { }
 
         public bool getStatus()
         {
